Rearrange files to work correctly with firefox extension security features
