+++ extension/TrawlerInfo.css
@@ -1,0 +1,3 @@
+#paragraph-view {
+    font-size: medium;
+}
