+++ extension/TrawlerInfo.html
@@ -1,0 +1,31 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <meta charset="utf-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+        <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+        <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+        <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+        <link rel="stylesheet" href="TrawlerInfo.css"
+              
+    </head>
+    <div class="container-fluid">
+        <div class="col-sm-9">
+                <h1>Trawler Anti-Phising Extension</h1>
+                <hr>
+                <p id="paragraph-view">This extension will help prevent Phising attempts <br>
+                    No cybersecurity software is perfect and neither is this extension, use caution when browsing the internet <br>
+                    Report bugs by creating a <a href=https://github.com/melbing1/Trawler>Github</a> issue.
+                 </p>
+                <br><br>
+                <hr>
+                </div>
+        </div>
+    </div>
+    </div>
+      
+      <footer class="container-fluid">
+        <p>Thank you for installing our extension!</p>
+      </footer>
+    </body>
+</html>
