+++ index.html
@@ -2,6 +2,7 @@
 <html lang="en">
     <head>
         <meta charset="utf-8"/>
+        <!-- <script src="extension/main.js"></script> -->
         Trawler Extension Testing
     </head>
     <body>
